+++ README.md
@@ -12,7 +12,9 @@ A beautiful way to manage your secrets in Vault
 Configuration is accessed by clicking on the configuration cog on the login page.
 ![Configuration](https://github.com/djenriquez/vault-ui/raw/master/images/Config.png)
 ### Vault Endpoint
-Users can enter in the full endpoint to Vault, including scheme.
+Users can enter in the full endpoint to Vault, including scheme.  When running the docker image, it is possible to
+set an environment variable `VAULT_DEFAULT_URL`, which will set the default endpoint for users without needing them
+to enter one in the UI.  This default can be overridden if the user fills out the endpoint manually.
 
 ## Authentication
 Currently supported authentication backends:
