+++ index.html
@@ -17,6 +17,7 @@
     </head>
     <body>
         <div id="app"></div>
+        <script>window.defaultUrl = "{{defaultUrl}}"</script>
         <script src="/assets/bundle.js"></script>
     </body>
 </html>
